FESolvers

Documentation for FESolvers.jl.

FESolvers.solve_problem! — Function

solve_problem!(problem, solver)

Solve a given user problem using the chosen solver

For details on the functions that should be defined for problem, see User problem

FESolvers.QuasiStaticSolver — Type

QuasiStaticSolver(nlsolver, timestepper)

A quasi-static solver that solves problems of the type R(X(t),t)=0. It has two parts: A nonlinear solver (see Nonlinear solvers) and a time stepper (see Time steppers).