Linear solvers

FESolvers.BackslashSolver — Type

BackslashSolver()

The standard julia linear solver using Δx .= -K\r

FESolvers.LinearSolveSolver — Type

LinearSolveSolver([alg], K, r=zeros(eltype(K), size(K,1)))

Create a linear solver with an algorithm alg from LinearSolve.jl. Please see LinearSolve.jl's documentation for different solver algorithms. If not given, LinearSolve's default algorithm will be used.

Extension

Using this solver requires using or importing LinearSolve.jl

Custom linear solver

A linear solver should support the solve_linear! function specified below.

FESolvers.solve_linear! — Function

solve_linear!(Δx, K, r, linearsolver)

Using the method specified by linearsolver, solve K Δx = -r for Δx