Nonlinear solvers

FESolvers.NewtonSolver — Type

NewtonSolver(;
    linsolver=BackslashSolver(), linesearch=NoLineSearch(), 
    maxiter=10, tolerance=1.e-6,
    update_jac_first=true, update_jac_each=true)

Use the standard Newton-Raphson solver to solve the nonlinear problem r(x) = 0 with tolerance within the maximum number of iterations maxiter.

Quasi-Newton methods

Linesearch: The linsolver argument determines the used linear solver whereas the linesearch can be set currently between NoLineSearch or ArmijoGoldstein. The latter globalizes the Newton strategy.

**Jacobian updates: ** The keyword update_jac_first decides if the jacobian from the previously converged time step should be updated after calling update_to_next_step!, or to use the old. Setting update_jac_each implies that the jacobian will not be updated during the iterations. If both update_jac_each and update_jac_first are false, the initial jacobian will be used throughout. Note that these keywords require that the problem respects the update_jacobian keyword given to update_problem!. For time-independent problems or time-depdent problems with constant time steps, update_jac_first=false is often a good choice. However, for time-dependent problems with changing time step length, the standard solver (default), may work better.

Missing docstring.

Missing docstring for AdaptiveNewtonSolver. Check Documenter's build log for details.

FESolvers.SteepestDescent — Type

SteepestDescent(;maxiter=10, tolerance=1.e-6)

Use a steepest descent solver to solve the nonlinear problem r(x) = 0, which minimizes a potential $\Pi$ with tolerance and the maximum number of iterations maxiter.

This method is second derivative free and is not as locally limited as a Newton-Raphson scheme. Thus, it is especially suited for strongly nonlinear behavior with potentially vanishing tangent stiffnesses. For this method, it is required to implement getdescentpreconditioner or alternatively getsystemmatrix with SteepestDescent.

FESolvers.LinearProblemSolver — Type

LinearProblemSolver(;linsolver=BackslashSolver())

This is a special type of "Nonlinear solver", which actually only solves linear problems, but allows all other features (i.e. time stepping and postprocessing) of the FESolvers package to be used. In particular, it allows you to maintain all other parts of your problem exactly the same as for a nonlinear problem, but it is possible to get better performance as it is, in principle, not necessary to assemble twice in each time step.

This solver is specialized for linear problems of the form

\[\boldsymbol{r}(\boldsymbol{x}(t),t)=\boldsymbol{K}(t) \boldsymbol{x}(t) - \boldsymbol{f}(t)\]

where $\boldsymbol{K}=\partial \boldsymbol{r}/\partial \boldsymbol{x}$. It expects that $\boldsymbol{x}(t)$ and $\boldsymbol{r}(t)$ have been updated to $\boldsymbol{x}_\mathrm{bc}$ and $\boldsymbol{r}_\mathrm{bc}=\boldsymbol{K}(t)\boldsymbol{x}_\mathrm{bc}-\boldsymbol{f}(t)$, such that

\[\boldsymbol{x}(t) = \boldsymbol{x}_\mathrm{bc} - \boldsymbol{K}^{-1}(t)\boldsymbol{r}_\mathrm{bc} = \boldsymbol{x}_\mathrm{bc} - \boldsymbol{K}^{-1}(t)\left[\boldsymbol{K}(t)\boldsymbol{x}_\mathrm{bc}-\boldsymbol{f}(t)\right] = \boldsymbol{K}^{-1}(t)\boldsymbol{f}(t)\]

is the solution to the current time step. This normally implies when using Ferrite the same procedure as for nonlinear problems, i.e. that the boundary conditions are applied in update_to_next_step! and update_problem!, as well as the calculation of the residual according to above and that apply_zero!(K,r,ch) is called (on both the residual and the stiffness matrix).

If you have strange results when running the LinearProblemSolver, please ensure that the problem converges in one iteration for the NewtonSolver

Problem update specification

An UpdateSpec, which can be querried for information, is passed to update_problem! to give instructions on how to update the problem.

FESolvers.UpdateSpec — Type

UpdateSpec(;jacobian, residual, type=nothing)

An UpdateSpec is sent to update_problem! to pass the problem information about how it should be updated. The following methods should be used by the problem to request information

should_update_jacobian(::UpdateSpec)::Bool: Self explanatory

should_update_residual(::UpdateSpec)::Bool: Self explanatory

get_update_type(::UpdateSpec): How should the problem be updated? This is used by special solvers where, for example, different approximations of the stiffness is available, see e.g. AdaptiveNewtonSolver

Custom solvers

A custom nonlinear solver can be written by tapping into the existing functions at different levels. For example, LinearProblemSolver defines a custom solve_nonlinear! that uses the default calculate_update!.

All nonlinear solvers are expected to implement the following methods,

FESolvers.getmaxiter — Function

getmaxiter(nlsolver)

Returns the maximum number of iterations allowed for the nonlinear solver

FESolvers.getnumiter — Function

getnumiter(nlsolver)

Returns the last number of iterations used by the nonlinear solver

FESolvers.get_convergence_measures — Function

get_convergence_measures(nlsolver[, i])

Get the convergence measures for iteration i from the nlsolver. If i is not given, the full history is returned. Note that unless i::Int is given, a new vector will be allocated.

FESolvers.gettolerance — Function

gettolerance(nlsolver)

Returns the iteration tolerance for the solver

FESolvers.get_initial_update_spec — Function

get_initial_update_spec(nlsolver)::UpdateSpec

How to update initially (before starting time stepping)

After implementing these, one can either implement solve_nonlinear!, or a set of method described below to use the default implementation.

FESolvers.solve_nonlinear! — Function

solve_nonlinear!(problem, nlsolver, last_converged)

Solve the current time step in the nonlinear problem, (r(x) = 0), by using the nonlinear solver nlsolver. last_converged::Bool is just for information if the last time step converged or not. In many cases it suffices to overload calculate_update! for a custom nonlinear solver.

Methods required by solve_nonlinear!

To use the default implementation of solve_nonlinear!, FESolvers.calculate_update! and the other methods in the list below must be implemented for the specific nonlinear solver. The default implementation of calculate_update! may be used as well, see the description below.

FESolvers.calculate_update! — Function

function calculate_update!(Δx, problem, nlsolver)

According to the nonlinear solver, nlsolver, at iteration iter, calculate the update, Δx to the unknowns x.

FESolvers.update_state! — Function

update_state!(nlsolver, problem, r)

A nonlinear solver may solve information about its convergence state. r is the output from calculate_convergence_measure when this function is called by the default implementation of check_convergence_criteria.

FESolvers.reset_state! — Function

reset_state!(nlsolver)

This function is used to reset nlsolver's state at the beginning of each new time step.

FESolvers.get_first_update_spec — Function

get_first_update_spec(nlsolver, last_converged::Bool)

Get the update specification for nlsolver in the first iteration of the time step. last_converged indicates if the previous time step converged, or if this is a retry using, e.g., a different time step.

FESolvers.get_update_spec — Function

get_update_spec(nlsolver)::UpdateSpec

Get the update specification during regular iterations. It is the nlsolver's job to keep track of any state required for deciding changes to the update specification during iterations.

FESolvers.maybe_reset_problem! — Function

maybe_reset_problem!(problem, Δa_old, nlsolver)

Note: Custom nonlinear solvers may rely on the no-op default implementation.

After checking for convergence, the default implementation of solve_nonlinear will call this function, allowing the nlsolver to call update_problem!(problem, -Δa) to reset the problem to the state at the last iteration (this is useful if e.g. the jacobian should be calculated differently, or if a higher dampening factor should be used when calculating the update).

Methods required by calculate_update!

To support the default calculate_update! implementation, the following methods must be implemented for the given solver.

FESolvers.get_linear_solver — Function

get_linear_solver(nlsolver)

Get the linear solver used by the nonlinear solver

FESolvers.get_linesearch — Function

get_linesearch(nlsolver)

Returns the used linesearch of the nonlinear solver.

Linesearch

Some nonlinear solvers can use linesearch as a complement, and the following linesearches are included.

FESolvers.NoLineSearch — Type

NoLineSearch()

Singleton that does not perform a linesearch when used in a nonlinear solver

FESolvers.ArmijoGoldstein — Type

Armijo-Goldstein{T}(;β=0.9,μ=0.01,τ0=1.0,τmin=1e-4)

Backtracking line search based on the Armijo-Goldstein condition

\[\Pi(\boldsymbol{u} + \tau \Delta\boldsymbol{u}) \leq \Pi(\boldsymbol{u}) - \mu\tau\delta\Pi(\boldsymbol{u})[\Delta \boldsymbol{u}]\]

where $\Pi$ is the potential, $\tau$ the stepsize, and $\delta\Pi$ the residuum.

#Fields

β::T = 0.9 constant factor that changes the steplength τ in each iteration

μ::T = 0.01 second constant factor that determines how much the potential needs to decrease additionally

τ0::T = 1.0 start stepsize

τmin::T = 1e-4 minimal stepsize

Custom linesearch

A custom linesearch should implement the following function

Missing docstring.

Missing docstring for linesearch!. Check Documenter's build log for details.