Time steppers

A time stepper should support the following functions

FESolvers.initial_time — Function

initial_time(timestepper)

Return the starting time for the given timestepper

FESolvers.islaststep — Function

islaststep(timestepper, time, step)->Bool

Return true if the current step/time is the last step, return false otherwise

FESolvers.update_time — Function

update_time(solver, time, step, converged::Bool)
update_time(timestepper, nlsolver, time, step, converged::Bool)

Return the next time and step number, depending on if the previous time step converged or not. If not converged, return the same step but a new_time<time to reduce the time step. If it is not possible to retry with shorter timestep, throw ConvergenceError. If converged, update time step as planned. Note: The full solver is given as input to allow specialization on e.g. if a Newton iteration has required many iterations, shorten the next time step as a precausionary step.

Note that a call to the first definition is forwarded to the second function definition by decomposing the solver, unless another specialization is defined.

Implemented steppers

FESolvers.FixedTimeStepper — Type

FixedTimeStepper(num_steps::int, Δt, t_start=0)
FixedTimeStepper(t::Vector)

A time stepper which gives fixed time steps. If the convenience interface is used, constant increments are used. Note that length(t)=num_steps+1 since the first value is just the initial value and is not an actual step.

FESolvers.AdaptiveTimeStepper — Type

AdaptiveTimeStepper(
    Δt_init::T, t_end::T; 
    t_start=zero(T), Δt_min=Δt_init, Δt_max=typemax(T), 
    change_factor=T(0.5), optiter_ratio=T(0.5), k=one(T)) where T

An adaptive time stepper with an initial step Δt_init and total time t_end. Two ways of adaption:

If the previous attempt did not converge, the time

step is reduced as Δt*=change_factor and the step is retried.

If convergence, the next time step depends on how many iterations was

required to converge; numiter. The time step is changed as Δt*=change_factor^(k*m), where m=(numiter-optiter)/(maxiter-optiter). In this expression, maxiter and optiter are the maximum and optimum number of iterations for the nonlinear solver. optiter=floor(maxiter*optiter_ratio) and maxiter is obtained from the nonlinear solver (via getmaxiter(s))

If numiter=maxiter, then m=1 and the time step update is the same as for a non-converged solution if k=1. Note that k>0, change_factor∈[0,1], and optiter_ratio∈[0,1] are expected, otherwise warnings are thrown.