FerriteSolvers

Documentation for FerriteSolvers.

FerriteSolvers.calculate_residualnorm

FerriteSolvers.getjacobian

FerriteSolvers.getresidual

FerriteSolvers.getunknowns

FerriteSolvers.handle_converged!

FerriteSolvers.postprocess!

FerriteSolvers.update_problem!

FerriteSolvers.update_to_next_step!

FerriteSolvers.calculate_residualnorm — Function

calculate_residualnorm(problem)

Calculate the value of the residual norm to be compared with the iteration tolerance

FerriteSolvers.getjacobian — Function

getjacobian(problem)

Return the current jacobian of the problem

FerriteSolvers.getresidual — Function

getresidual(problem)

Return the current residual vector of the problem

FerriteSolvers.getunknowns — Function

getunknowns(problem)

Return the current vector of unknown values

FerriteSolvers.handle_converged! — Function

handle_converged!(problem)

Do necessary update operations once it is known that the problem has converged. E.g., update old values to the current. Only called directly after the problem has converged.

FerriteSolvers.postprocess! — Function

postprocess!(problem, step)

Perform any postprocessing at the current time and step nr step Called after time step converged, and after handle_converged!

FerriteSolvers.update_problem! — Function

update_problem!(problem, Δa=0*getdofvector(problem))

Apply zero change to Dirichlet conditions on Δa Assemble the stiffness and residual.

FerriteSolvers.update_to_next_step! — Function

update_to_next_step!(problem, time)

Update prescribed values, external loads etc. for the given time Called in the beginning of each new time step. Note: For adaptive time stepping, it may be called with a lower time than the previous time if the solution did not converge.