FerriteSolvers

Documentation for FerriteSolvers.jl.

FerriteSolvers.solve_ferrite_problem! — Function

solve_ferrite_problem!(solver, problem)

Solve a time-dependent problem r(x(t),t)=0 for x(t), stepping throught the time t, using the solver. The following functions must be defined for the user-defined problem:

getunknowns(problem)

getresidual(problem)

getjacobian(problem)

update_to_next_step!(problem, t)

update_problem!(problem, Δx)

calculate_convergence_criterion(problem)

handle_converged!(problem)

Additionally, one can define postprocess!(problem, step) For details on each function above, please see the respective function's documentation under User problem

FerriteSolvers.FerriteSolver — Type

FerriteSolver(nlsolver, timestepper, linsolver=BackslashSolver())

The standard solver, with two parts: A nonlinear solver (see Nonlinear solvers) and a time stepper (see Time steppers).