Linear solvers

A linear solver should support the solve_linear! function specified below.

FESolvers.solve_linear! — Function

solve_linear!(Δx, K, r, linearsolver)

Using the method specified by linearsolver, solve K Δx = -r for Δx

Implemented Solvers

BackslashSolver

FESolvers.BackslashSolver — Type

BackslashSolver()

The standard julia linear solver using Δx .= -K\r

LinearSolve.jl

The linear solvers in LinearSolve.jl are available if the LinearSolve.jl package is available (implemented via Requires.jl). This also includes their default solver that is supplied setting the linear solver to nothing. Please see LinearSolve.jl's documentation for their different solvers.