Nonlinear solvers

A nonlinear solver should support the solve_nonlinear! function specified below.

FerriteSolvers.solve_nonlinear! — Function

solve_nonlinear!(solver::FerriteSolver{<:NLS}, problem)

Solve one step in the nonlinear problem, given as r(x) = 0, by using the solver of type NLS.

Implemented Solvers

NewtonSolver

FerriteSolvers.NewtonSolver — Type

NewtonSolver(;maxiter=10, tolerance=1.e-6)

Use the standard NewtonRaphson solver to solve the nonlinear problem r(x) = 0 with tolerance within the maximum number of iterations maxiter