Nonlinear solvers

A nonlinear solver should support the solve_nonlinear! function specified below.

FESolvers.solve_nonlinear! — Function

solve_nonlinear!(problem, nlsolver)

Solve the current time step in the nonlinear problem, (r(x) = 0), by using the nonlinear solver: nlsolver.

It can do so, by supporting the following functions

FESolvers.calculate_update! — Function

function calculate_update!(Δx, problem, nlsolver, iter)

According to the nonlinear solver, nlsolver, at iteration iter, calculate the update, Δx to the unknowns x.

FESolvers.getmaxiter — Function

getmaxiter(nlsolver)

Returns the maximum number of iterations allowed for the nonlinear solver

FESolvers.gettolerance — Function

gettolerance(nlsolver)

Returns the iteration tolerance for the solver

and optionally

FESolvers.update_state! — Function

update_state!(nlsolver, r)

A nonlinear solver may solve information about its convergence state. r is the output from calculate_convergence_measure when this function is called by the default implementation of check_convergence_criteria. update_state! is optional to implement

FESolvers.reset_state! — Function

reset_state!(nlsolver)

If update_state! is implemented, this function is used to reset its state at the beginning of each new time step.

The implemented solvers additionally support

FESolvers.getnumiter — Function

getnumiter(nlsolver)

Returns the last number of iterations used by the nonlinear solver

FESolvers.getlinesearch — Function

getlinesearch(nlsolver)

Returns the used linesearch of the nonlinear solver.

Implemented Solvers

FESolvers.LinearProblemSolver — Type

LinearProblemSolver(;linsolver=BackslashSolver())

This is a special type of "Nonlinear solver", which actually only solves linear problems, but allows all other features (i.e. time stepping and postprocessing) of the FESolvers package to be used. In particular, it allows you to maintain all other parts of your problem exactly the same as for a nonlinear problem, but it is possible to get better performance as it is, in principle, not necessary to assemble twice in each time step.

This solver is specialized for linear problems of the form

\[\boldsymbol{r}(\boldsymbol{x}(t),t)=\boldsymbol{K}(t) \boldsymbol{x}(t) - \boldsymbol{f}(t)\]

where $\boldsymbol{K}=\partial \boldsymbol{r}/\partial \boldsymbol{x}$. It expects that $\boldsymbol{x}(t)$ and $\boldsymbol{r}(t)$ have been updated to $\boldsymbol{x}_\mathrm{bc}$ and $\boldsymbol{r}_\mathrm{bc}=\boldsymbol{K}(t)\boldsymbol{x}_\mathrm{bc}-\boldsymbol{f}(t)$, such that

\[\boldsymbol{x}(t) = \boldsymbol{x}_\mathrm{bc} - \boldsymbol{K}^{-1}(t)\boldsymbol{r}_\mathrm{bc} = \boldsymbol{x}_\mathrm{bc} - \boldsymbol{K}^{-1}(t)\left[\boldsymbol{K}(t)\boldsymbol{x}_\mathrm{bc}-\boldsymbol{f}(t)\right] = \boldsymbol{K}^{-1}(t)\boldsymbol{f}(t)\]

is the solution to the current time step. This normally implies when using Ferrite the same procedure as for nonlinear problems, i.e. that the boundary conditions are applied in update_to_next_step! and update_problem!, as well as the calculation of the residual according to above and that apply_zero!(K,r,ch) is called (on both the residual and the stiffness matrix).

If you have strange results when running the LinearProblemSolver, please ensure that the problem converges in one iteration for the NewtonSolver

FESolvers.NewtonSolver — Type

NewtonSolver(;linsolver=BackslashSolver(), linesearch=NoLineSearch(), maxiter=10, tolerance=1.e-6)

Use the standard NewtonRaphson solver to solve the nonlinear problem r(x) = 0 with tolerance within the maximum number of iterations maxiter. The linsolver argument determines the used linear solver whereas the linesearch can be set currently between NoLineSearch or ArmijoGoldstein. The latter globalizes the Newton strategy.

FESolvers.SteepestDescent — Type

SteepestDescent(;maxiter=10, tolerance=1.e-6)

Use a steepest descent solver to solve the nonlinear problem r(x) = 0, which minimizes a potential \Pi with tolerance and the maximum number of iterations maxiter.

This method is second derivative free and is not as locally limited as a Newton-Raphson scheme. Thus, it is especially suited for stronlgy nonlinear behavior with potentially vanishing tangent stiffnesses.

Linesearch

Some nonlinear solvers can use linesearch as a complement, and the following linesearches are included.

FESolvers.NoLineSearch — Type

Singleton that does not perform a linesearch when used in a nonlinear solver

FESolvers.ArmijoGoldstein — Type

Armijo-Goldstein{T}(;β=0.9,μ=0.01,τ0=1.0,τmin=1e-4)

Backtracking line search based on the Armijo-Goldstein condition

\[\Pi(\boldsymbol{u} + \tau \Delta\boldsymbol{u}) \leq \Pi(\boldsymbol{u}) - \mu\tau\delta\Pi(\boldsymbol{u})[\Delta \boldsymbol{u}]\]

where $\Pi$ is the potential, $\tau$ the stepsize, and $\delta\Pi$ the residuum.

#Fields

β::T = 0.9 constant factor that changes the steplength τ in each iteration

μ::T = 0.01 second constant factor that determines how much the potential needs to decrease additionally

τ0::T = 1.0 start stepsize

τmin::T = 1e-4 minimal stepsize