Nonlinear solvers

A nonlinear solver should support the solve_nonlinear! function specified below.

FerriteSolvers.solve_nonlinear! — Function

solve_nonlinear!(nlsolver, problem)

Solve the current time step in the nonlinear problem, (r(x) = 0), by using the nonlinear solver: nlsolver.

It can do so, by supporting the following functions

FerriteSolvers.calculate_update — Function

function calculate_update(problem, nlsolver, iter)

According to the nonlinear solver, nlsolver, at iteration iter, calculate the update, Δx to the unknowns x.

FerriteSolvers.getmaxiter — Function

getmaxiter(nlsolver)

Returns the maximum number of iterations allowed for the nonlinear solver

FerriteSolvers.gettolerance — Function

gettolerance(nlsolver)

Returns the iteration tolerance for the solver

and optionally

FerriteSolvers.update_state! — Function

update_state!(nlsolver, r)

A nonlinear solver may solve information about its convergence state. r is the output from calculate_convergence_measure when this function is called by the default implementation of check_convergence_criteria. update_state! is optional to implement

FerriteSolvers.reset_state! — Function

reset_state!(nlsolver)

If update_state! is implemented, this function is used to reset its state at the beginning of each new time step.

Implemented Solvers

FerriteSolvers.NewtonSolver — Type

NewtonSolver(;linsolver=BackslashSolver(), linesearch=NoLineSearch(), maxiter=10, tolerance=1.e-6)

Use the standard NewtonRaphson solver to solve the nonlinear problem r(x) = 0 with tolerance within the maximum number of iterations maxiter. The linsolver argument determines the used linear solver whereas the linesearch can be set currently between NoLineSearch or ArmijoGoldstein. The latter globalizes the Newton strategy.

FerriteSolvers.SteepestDescent — Type

SteepestDescent(;maxiter=10, tolerance=1.e-6)

Use a steepest descent solver to solve the nonlinear problem r(x) = 0, which minimizes a potential \Pi with tolerance and the maximum number of iterations maxiter.

This method is second derivative free and is not as locally limited as a Newton-Raphson scheme. Thus, it is especially suited for stronlgy nonlinear behavior with potentially vanishing tangent stiffnesses.

Linesearch

Some nonlinear solvers can use linesearch as a compliment, and the following linesearches are included.

FerriteSolvers.NoLineSearch — Type

Singleton that does not perform a linesearch when used in a nonlinear solver

FerriteSolvers.ArmijoGoldstein — Type

Armijo-Goldstein{T}(;β=0.9,μ=0.01,τ0=1.0,τmin=1e-4)

Backtracking line search based on the Armijo-Goldstein condition

\[\Pi(\boldsymbol{u} + \tau \Delta\boldsymbol{u}) \leq \Pi(\boldsymbol{u}) - \mu\tau\delta\Pi(\boldsymbol{u})[\Delta \boldsymbol{u}]\]

where $\Pi$ is the potential, $\tau$ the stepsize, and $\delta\Pi$ the residuum.

#Fields

β::T = 0.9 constant factor that changes the steplength τ in each iteration

μ::T = 0.01 second constant factor that determines how much the potential needs to decrease additionally

τ0::T = 1.0 start stepsize

τmin::T = 1e-4 minimal stepsize