User problem

The key to using the FerriteSolvers.jl package is to define your problem. This problem should support the following functions in order for the solver to solve your problem.

FerriteSolvers.getunknowns — Function

getunknowns(problem)

Return the current vector of unknown values

FerriteSolvers.getresidual — Function

getresidual(problem)

Return the current residual vector of the problem

FerriteSolvers.getjacobian — Function

getjacobian(problem)

Return the current jacobian of the problem

FerriteSolvers.update_to_next_step! — Function

update_to_next_step!(problem, time)

Update prescribed values, external loads etc. for the given time Called in the beginning of each new time step. Note: For adaptive time stepping, it may be called with a lower time than the previous time if the solution did not converge.

FerriteSolvers.update_problem! — Function

update_problem!(problem, Δx=0*getunknowns(problem))

Assemble the residual and stiffness for x+=Δx.

Some linear solvers may be inaccurate, and if modified stiffness is used to enforce constraints on x, it is good the force Δx=0 on these components inside this function.

Note that the function must also support only one argument: problem, this version is called the first time after update-update_to_next_step! and should default to Δx=0

FerriteSolvers.calculate_residualnorm — Function

calculate_residualnorm(problem)

Calculate the value of the residual norm to be compared with the iteration tolerance

FerriteSolvers.handle_converged! — Function

handle_converged!(problem)

Do necessary update operations once it is known that the problem has converged. E.g., update old values to the current. Only called directly after the problem has converged.

FerriteSolvers.postprocess! — Function

postprocess!(problem, step)

Perform any postprocessing at the current time and step nr step Called after time step converged, and after handle_converged!