User problem

The key to using the FerriteSolvers.jl package is to define your problem. This problem should support a set of functions in order for the solver to solve your problem. While some functions are always required, some are only required by certain solvers. Furthermore, a two-level API exist: Simple and advanced. The simple API does not expose which solver is used, while the advanced API requires you to dispatch on the type of solver.

Applicable to all solvers

FerriteSolvers.getunknowns — Function

getunknowns(problem)

Return the current vector of unknown values

FerriteSolvers.getresidual — Function

getresidual(problem)

Return the current residual vector of the problem

FerriteSolvers.update_to_next_step! — Function

update_to_next_step!(problem, time)

Update prescribed values, external loads etc. for the given time.

This function is called in the beginning of each new time step. Note that for adaptive time stepping, it may be called with a lower time than the previous time if the solution did not converge.

FerriteSolvers.update_problem! — Function

update_problem!(problem)
update_problem!(problem, Δx)

Assemble the residual and stiffness for x+=Δx.

Some linear solvers may be inaccurate, and if modified stiffness is used to enforce constraints on x, it is good the force Δx=0 on these components inside this function.

Δx is not given in the first call after update_to_next_step! in which case no change of x should be made.

FerriteSolvers.handle_converged! — Function

handle_converged!(problem)

Do necessary update operations once it is known that the problem has converged. E.g., update old values to the current. Only called directly after the problem has converged.

FerriteSolvers.postprocess! — Function

postprocess!(problem, step)

Perform any postprocessing at the current time and step nr step Called after time step converged, and after handle_converged!

Simple API

FerriteSolvers.calculate_convergence_measure — Function

calculate_convergence_measure(problem) -> AbstractFloat

Calculate a value to be compared with the tolerance of the nonlinear solver. A standard case when using Ferrite.jl is norm(getresidual(problem)[Ferrite.free_dofs(dbcs)]) where dbcs::Ferrite.ConstraintHandler.

The advanced API alternative is check_convergence_criteria

FerriteSolvers.getjacobian — Function

getjacobian(problem)

Return the jacobian drdx, or approximations thereof.

Must be defined for NewtonSolver, but can also be defined by the advanced API alternative getsystemmatrix: getsystemmatrix(problem, ::NewtonSolver)

FerriteSolvers.getdescentpreconditioner — Function

getdescentpreconditioner(problem)

Return a preconditioner K for calculating the descent direction p, considering solving r(x)=0 as a minimization problem of f(x) where r=∇f. The descent direction is then p = K⁻¹ ∇f

Used by the SteepestDescent solver, and defaults to I if not defined. The advanced API alternative is getsystemmatrix: getsystemmatrix(problem, ::SteepestDescent)

Advanced API

FerriteSolvers.getsystemmatrix — Function

getsystemmatrix(problem,nlsolver)

Return the system matrix of the problem. For a Newton solver this method should return the Jacobian, while for a steepest descent method this can be a preconditioner as e.g., the L2 product of the gradients. By default the system matrix for the SteepestDescent method is the unity matrix and thus, renders a vanilla gradient descent solver.

FerriteSolvers.check_convergence_criteria — Function

check_convergence_criteria(problem, nlsolver) -> Bool

Check if problem has converged and update the state of nlsolver wrt. number of iterations and a convergence measure if applicable.